auth routes

+++ apps/server/routes/expenses.ts
@@ -5,7 +5,7 @@ import { z } from 'zod'
 const expenseSchema = z.object({
 	id: z.number().int().positive().min(1),
 	title: z.string().min(3).max(100),
-	amount: z.number().int().positive(),
+	amount: z.number().positive(),
 })
 
 const createPostSchema = expenseSchema.omit({ id: true })
